Related work items: #4396, #12947, #12948, #12950, #12951, #13098, #13099, #13100, #13101, #13104, #13106, #13107, #13108, #13109

+++ java/commoncomponent/common.serialization.deserialization/src/main/java/net/atos/daf/ct2/pojo/standard/ROEventData.java
@@ -36,7 +36,7 @@ public class ROEventData  implements Serializable {
     @JsonProperty(value = "Message_Timestamp")
     private Date messageTimestamp;
     @JsonProperty(value = "Message_NumSeq")
-    private Integer messageNumSeq;
+    private Long messageNumSeq;
     @JsonProperty(value = "Message_Name")
     private String messageName;
